Fixes #8 Hotmail email mailbox_validator is incorrect
It seems to require a start of a letter, then any letter, digit, _ -
and period. Though the full stop (period) does not seem to able to end
the name, other punctuation can, and can even occur in succession
(e.g. "w.-asdf----") I have changed the regular expression to match this.