Fixes yahoo.badtld check, valid IP return value
I was giving "yahoo.*" a pass since it is well-known and and had
multiple TLD's. Any bad TLD still passed, so this "skip" was removed.
(Probably also present for hotmail.*, etc.)

Also the return from host.valid_ip? was exected to be true, not just
truthy, so it now returns a boolean.