Updates Commit 17bc577 for Missing MX Hosts
Interesting:

    EmailAddress.new("a@yahoo.wtf").host.exchangers.mxers

yahoo.wtf has an empty hostname in the MX record, which is wierd.

    yahoo.wtf.		804	IN	MX	0 .

Looking up the IP address for an empty hostname returned:

  IPSocket::getaddress("")
                            #=> "0.0.0.0"   # MRI
                            #=> "127.0.0.1" # Jruby

This caused Jruby run to fail in Travis CI. While they return
different values, I would lean towards 0.0.0.0 being more correct
(unassigned IP) per RFC3300, and not resolving to the loopback address.

That said, a valid hostname is required for an MX record. If an empty or
unresolvable hostname is returned, it shouldn't be included in the
result set of MX IP's. (Invalid/Unknown ones raise a SocketError and have
been excluded already.)

Host validation requires at least 1 MX IP (for the dns_lookup=:mx rule)
and should still work as expected. Badly configured domains shouldn't
be considered valid. This *could* be a temporary condition, but what
should we do here in that case?