Adds network connectivity recovery logic
We can't tell (in Ruby?) that the network is down, so if a dns lookup
returned a SocketError, it can mean either not found or network
unavailable. Until I find a better way, I test the network on that
exception by looking up "example.com". If it works, then assume it was a
not-found condition. For unavailable, a dummy record is returned.

Also, moved the a-record lookup into the same exchanger object, with the
same network test.

Exchanger objects are cached. When I pull it out of the cache, I check
to see if the network-down occurred, and if so, throw it away and create
a new one so it can work properly when the network becomes available.