Fixes #53 host_validation: :syntax using DNS
There were two settings, this and dns_lookup: :off which isn't the best
way to configure it. Changed to infer dns_lookup :off on host_validation
:syntax.

While host_validation :syntax was not performing DNS to validate, it
still used it to lookup the email provider. That will now honor DNS
settings as well.