Fixes #58 reworks regex.match for older rubies
A change from
   r =~ /.../
to
   /.../.match?(r)
broke support for older Rubyies < 2.3.7 (earliest version tested with
travis). While these versions are End Of Life and we don't target
support for them, compatible syntax changes should be made if possible.
This changes above syntax to
    r.match?(/.../)

Note this was an automated change performed by Rubocop.