Fixes #60 for Ruby <=2.3.0, Release v0.1.16
This is a compatability release for earlier, EOL versions of Ruby
that are not supported but can still operate on earlier versions.

After March 2020 (today), Ruby 2.4.x is EOL. So long, old friend!

Travis-CI builds only work on Ruby >=2.3.0. I can't give assurance
of earlier rubies continuing to work.

While earlier rubies this may work, but not probably not with the
Rails dependencies here. These are development dependencies only,
so won't prevent inclusion in your Bundler. If you don't use the
Rails EmailAddressType's, you should be good.