Fixes #62 GSuite hosted domain rules !== gmail.com
This breaks an original assumption that gmail.com address rules would be
the same for other hosted domains. Since this issue points out is not
true, this patch removes the connection to the address rules when
matched by exchanger domain name.

Additionally, the email hosting service can be found through the
exchanger, with a helper in the host object

    host.hosted_provider