Disables JRuby ActiveRecord checks (Bad CI Errors)
Hopefully someone can help out on this. The crash happens when it tries
to create the Sqlite table to test. Versions have changed a lot since
this was initially coded. JRuby can't run under openjdk, and I'm on an
M1 ARM architecture computer. Then my experience with Java is also
outdated is seems.