Fixes #91 Gmail address local size now 3..64 chars
It does seem that gmail addresses in the wild are less than 6 characters
even though it requires 6 characters in signup (even when I signed up by
invite back in beta). I'm dropping this down to 3 characters minimum to
see if that solves the problems. Bad addresses would be discovered by
bounced back emails.